batch: Extract baseline correspondence
The batch merge layer owns line-sequence merge and discard operations. It
carries the baseline correspondence records used by discard restoration.

That leaves discard-specific restoration mapping mixed with merge placement
and realized-entry traversal, so readers must inspect unrelated merge logic
before they can follow baseline restoration.

This commit moves the baseline region records and builder into
batch.baseline_correspondence, then imports the builder from merge.py for
the discard path.

The next commit will retarget the baseline correspondence tests to the new
module boundary.
